# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Tasks Management API",
        "contact": {
            "email": "malinowski.rad@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/send-notifications/tasks": {
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Prepare and sends all notifications",
                "operationId": "send-task-notifications",
                "responses": {
                    "204": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Display a listing of the resource.",
                "operationId": "index",
                "parameters": [
                    {
                        "name": "completed",
                        "in": "query",
                        "description": "Filter by completed flag",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "due_date[from]",
                        "in": "query",
                        "description": "Filter by due_date",
                        "schema": {
                            "type": "string",
                            "example": "2023-02-01"
                        }
                    },
                    {
                        "name": "due_date[to]",
                        "in": "query",
                        "description": "Filter by due_date",
                        "schema": {
                            "type": "string",
                            "example": "2025-02-01"
                        }
                    },
                    {
                        "name": "order_by[completed]",
                        "in": "query",
                        "description": "Sort by completed",
                        "schema": {
                            "type": "enum",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "order_by[due_date]",
                        "in": "query",
                        "description": "Sort by due_date",
                        "schema": {
                            "type": "enum",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "order_by[activities]",
                        "in": "query",
                        "description": "Sort by activities",
                        "schema": {
                            "type": "enum",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TaskResourceForCollection"
                                            }
                                        },
                                        "links": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Links"
                                            }
                                        },
                                        "meta": {
                                            "type": "object",
                                            "example": "{}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request. Validation or bussiness logic error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Store a newly created resource in storage.",
                "operationId": "store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreTaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Task"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request. Validation or bussiness logic error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tasks/{taskId}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Display the specified resource.",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "taskId",
                        "in": "path",
                        "description": "Task uniqal ID",
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/TaskResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request. Validation or bussiness logic error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Remove the specified resource from storage.",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "taskId",
                        "in": "path",
                        "description": "Task ID",
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "204": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request. Validation or bussiness logic error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Update the specified resource in storage.",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "taskId",
                        "in": "path",
                        "description": "Task ID",
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Task"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request. Validation or bussiness logic error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "StoreTaskRequest": {
                "required": [
                    "title"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Example task title"
                    },
                    "description": {
                        "type": "string",
                        "example": "Example task title",
                        "nullable": true
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-02-25",
                        "nullable": true
                    }
                },
                "type": "object",
                "xml": {
                    "name": "StoreTask"
                }
            },
            "UpdateTaskRequest": {
                "required": [
                    "title",
                    "completed"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Example task title"
                    },
                    "description": {
                        "type": "string",
                        "example": "Example task title",
                        "nullable": true
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-02-25",
                        "nullable": true
                    },
                    "completed": {
                        "type": "boolean",
                        "example": true,
                        "nullable": false
                    }
                },
                "type": "object",
                "xml": {
                    "name": "StoreTask"
                }
            },
            "Links": {
                "properties": {
                    "first": {
                        "type": "string",
                        "example": "http://example.com/api/element?page=1"
                    },
                    "last": {
                        "type": "string",
                        "example": "http://example.com/api/element?page=10"
                    },
                    "next": {
                        "type": "string",
                        "format": "http://example.com/api/element?page=5"
                    },
                    "prev": {
                        "type": "string",
                        "example": "http://example.com/api/element?page=3"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Links"
                }
            },
            "TaskResource": {
                "required": [
                    "id",
                    "title",
                    "completed"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Example task title"
                    },
                    "description": {
                        "type": "string",
                        "example": "Example task title",
                        "nullable": true
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-02-25",
                        "nullable": true
                    },
                    "completed": {
                        "type": "boolean",
                        "example": true,
                        "nullable": false
                    },
                    "activities": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    }
                },
                "type": "object",
                "xml": {
                    "name": "TaskResource"
                }
            },
            "TaskResourceForCollection": {
                "required": [
                    "id",
                    "title",
                    "completed"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Example task title"
                    },
                    "description": {
                        "type": "string",
                        "example": "Example task title",
                        "nullable": true
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-02-25",
                        "nullable": true
                    },
                    "completed": {
                        "type": "boolean",
                        "example": true,
                        "nullable": false
                    }
                },
                "type": "object",
                "xml": {
                    "name": "TaskResourceForCollection"
                }
            },
            "BaseModel": {
                "required": [
                    "title",
                    "completed"
                ],
                "properties": {
                    "created_at": {
                        "description": "Initial creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "deleted_at": {
                        "description": "Soft delete timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
            "Task": {
                "required": [
                    "title",
                    "completed"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "readOnly": true,
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Example task title"
                    },
                    "description": {
                        "type": "string",
                        "example": "Example task description",
                        "nullable": true
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-02-25",
                        "nullable": true
                    },
                    "completed": {
                        "type": "boolean",
                        "example": true,
                        "nullable": false
                    },
                    "created_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/created_at"
                    },
                    "updated_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/updated_at"
                    },
                    "deleted_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/deleted_at"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Task"
                }
            }
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://127.0.0.1:8000/oauth/authorize",
                        "tokenUrl": "http://127.0.0.1:8000/oauth/token",
                        "refreshUrl": "http://127.0.0.1:8000/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Tasks",
            "description": "Everything about your Tasks"
        }
    ]
}